+++ NEWS.md
@@ -1,8 +1,17 @@
 # qol 1.3.4 - DEVELOPMENT
 
+### New functionality
+
+* `load_file()`, `load_file_multi()`: When passing a named vector or list into the `keep` parameter, the original variables will directly be renamed. (18.07.2026)
+
+### Removed
+
+* `get_integer_length()`: There is no real benefit in this function. (08.07.2026)
+
 ### Fixed
 
 * `any_table()`: Variables generated with the `compute` parameter now receive a default `statistics` extension, if they are missing one, so that the column header is generated correctly and a number format is applied. (17.07.2026)
+* `mark_case()`: When marking `last` cases and the last expression within the `by` variable was NA, then the last case in the second to last expression became NA instead of TRUE. This is fixed now. (18.07.2026)
 
 # qol 1.3.3